Added /files page and global header.
* Added a /files page that shows all files. For now, this will just show
a bunch of sample files. Ofcourse, that will be changed later.

* The global header has the title, along with buttons to see the files
page, a login/logout button, and a theme switcher icon. For the logout
button, currently the website will override login procedure and log you
in. DO NOT click the logout button. It will lock you out of the files
pages and theres no way to log in again (except a hard refresh on the
page). The theme switcher button can remain, but its main purpose now is
to show you what the base themes are and you can view all and pick one.
My personal favorites are light and swiss. Though, I would like to
switch up the dark mode a bit so its less jarring and more subtle. I'm
of the opinion that we can use Apples design language manuals and use
the colors from there.

* The login/logout procedures are not well established yet but the
/src/contexts/AuthContext.js module gives a rough overview. The process
is most likely:

	* Check for existing login via cookies. Also need to query
	Google to confirm that cached OAuth token is valid. This is also
	where we would perform the check for mpw validity using the
	MANIFEST on drive.

	* If everything is fine, log in, otherwise, set the user state
	to null

	* Also, whenever we are doing any of the login/logout work set
	the loading state to true. This will ensure that no pages are
	rendered when we are unsure of auth status. An example of when
	this could cause an issue is if the user goes straight to the
	/files page and is not logged in. If the page tries to render
	when we don't know the auth status, we would get a bunch of
	NPE's and other nasty errors.

	TLDR: setLoading(true) -> Do a bunch of global auth stuff ->
	setLoading(false)