Overhaul of login page and mechanics.
Authcontext will now maintain 2 state vars: currentUser and
currentOAuthUser. Both will initally be `null'. On page load, the google
API will attempt to load the cached token, and log the user into Google
if it exists & is valid. This will set the `currentOAuthUser` state.

The `currentUser' state will be set on masterpassword authentication
now.

In order to log out of the Google Account, there are 2 ways:
* If the user is fully logged in, clicking the logout button will
log them out of both google and the masterpassword. We can debate
whether we still want them logged into Google though.
* If the user is partially logged in (only to OAuth), they can go to the
login page, and click the text that allows them to do this. This also
lets them swap accounts.